+++ language/move-core/types/Cargo.toml
@@ -19,6 +19,7 @@ mirai-annotations = "1.8.0"
 proptest-derive = { version = "0.1.2", default-features = false, optional = true }
 ref-cast = "1.0"
 serde = { version = "1.0.110", default-features = false }
+thiserror = "1.0"
 
 lcs = { path = "../../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-workspace-hack = { path = "../../../common/workspace-hack", version = "0.1.0" }
